+++ src/test/java/com/library/runner/CukesRunner.java
@@ -15,7 +15,7 @@ import org.junit.runner.RunWith;
         features = "src/test/resources/features" ,
         glue = "com/library/steps",
         dryRun = false,
-        tags = "@US01BHA"
+        tags = ""
 )
 public class CukesRunner {
