+++ src/test/java/com/library/steps/Hooks.java
@@ -13,7 +13,7 @@ import java.util.concurrent.TimeUnit;
 
 public class Hooks {
 
-    @Before("ui")
+    @Before("@ui")
     public void setUp(){
         System.out.println("this is coming from BEFORE");
         Driver.getDriver().manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
@@ -23,7 +23,7 @@ public class Hooks {
 
     }
 
-    @After("ui")
+    @After("@ui")
     public void tearDown(Scenario scenario){
         System.out.println("this is coming from AFTER");
 
@@ -36,12 +36,12 @@ public class Hooks {
 
     }
 
-    @Before("db")
+    @Before("@db")
     public void setUpDB(){
         DB_Util.createConnection();
     }
 
-    @After("db")
+    @After("@db")
     public  void destroyDB(){
         DB_Util.destroy();
