+++ src/test/resources/features/us01.feature
@@ -3,12 +3,12 @@
 Feature: As a data consumer, I want the user information are stored in mySql DB correctly in users table.
   Background:
     Given Establish the database connection
-   @db
+   @ui @db
   Scenario: verify users has unique IDs
     When Execute query to get all IDs from users
     Then verify all users has unique ID
 
-  @db
+  @ui @db
   Scenario: verify users table columns
     When Execute query to get all columns
     Then verify the below columns are listed in result
